% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWGmutations.R
\name{pcawg_load}
\alias{pcawg_load}
\title{Load PCAWG mafs}
\usage{
pcawg_load(cohort, verbose = TRUE)
}
\arguments{
\item{cohort}{abbreviation of PCAWG project. See \code{\link[=pcawg_available]{pcawg_available()}} for valid values (string)}

\item{verbose}{verbose (flag)}
}
\value{
MAF object compatible with maftools
}
\description{
Load PCAWG maf objects into R. Streams data from \href{https://github.com/CCICB/PCAWGmutations}{PCAWGmutationsDB} repo
}
\examples{
pcawg_load("Biliary-AdenoCA")

}
